Refactor/Minor: End files with newlines and strip trailing spaces
The former is required by some Unix tools to ensure that the tool processes the contents of the file correctly.  The latter (a) makes text-editing simpler by removing meaning-free whitespace characters that can only interfere with navigation, and (b) not generating huge amounts of diffs on every save for those (like me) whose text editors automatically strip the trailing whitespace.